working arte_to_tt, output is accepted by spikeparms and xclust3.  Though timestamps appear to be a little messed up (timecourses of amplitude look good but some spikeanal calls complain about repeated timestamps..  may be a problem with my spike detector double-counting spikes?

+++ src/extract/arte_to_tt.h
@@ -8,9 +8,9 @@
 #include "../global_defs.h"
 
 void write_spike(spike_net_t *spike);
-void init_filenames(int argc, char *argv[], char input_filename[], char output_filename[], uint16_t *trodename);
+void init_filenames(int argc, char *argv[], char input_filename[], char output_filename[], int *trodename);
 void write_file_header(spike_net_t *spike, int argc, char *argv[]);
-bool get_next_spike(spike_net_t *spike, uint16_t trodename);
+bool get_next_spike(spike_net_t *spike, int trodename);
 //void tfprintf(FILE *fp, char format_str[], ...);
 //void ttfprintf(FILE *fp, char format_str[], ...);
 void ps();
